display users for demo
