+++ nextjs-task-system/app/actions/getUsers.ts
@@ -1,0 +1,18 @@
+import prisma from "@/lib/prisma";
+
+export async function getUsers() {
+  try {
+    const users = await prisma.user.findMany({
+      select: {
+        id: true,
+        name: true,
+        email: true,
+        role: true,
+      },
+    });
+    return users;
+  } catch (error) {
+    console.error("Error fetching users:", error);
+    return [];
+  }
+}
