+++ nextjs-task-system/app/auth/signin/SignIn.tsx
@@ -1,0 +1,94 @@
+"use client";
+
+import { signIn } from "next-auth/react";
+import { ReactNode, useState } from "react";
+import { Button, TextInput, Alert, Card, Label } from "flowbite-react";
+import { useRouter } from "next/navigation";
+
+interface SignInProps {
+  children: ReactNode;
+}
+
+export const SignIn = ({ children }: SignInProps) => {
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [error, setError] = useState("");
+  const router = useRouter();
+
+  const handleSignIn = async (e: React.FormEvent) => {
+    try {
+      e.preventDefault();
+      setError("");
+      const result = await signIn("credentials", {
+        email,
+        password,
+        role: "Admin",
+        redirect: false,
+      });
+
+      if (result?.error) {
+        setError("Invalid credentials");
+      } else {
+        router.push("/");
+      }
+    } catch (error) {
+      setError("unexpected error");
+    }
+  };
+
+  return (
+    <div className="flex flex-1 items-center justify-center">
+      <Card className="w-full max-w-md rounded-lg p-6 shadow-md">
+        <h1 className="mb-6 text-center text-2xl font-bold dark:text-gray-200">
+          Sign In
+        </h1>
+        {children}
+        {error && (
+          <Alert color="failure" className="mb-4">
+            {error}
+          </Alert>
+        )}
+        <form onSubmit={handleSignIn} className="space-y-4">
+          <div>
+            <Label htmlFor="email" className="mb-1 block text-sm font-medium">
+              Email
+            </Label>
+            <TextInput
+              id="email"
+              type="email"
+              placeholder="Enter your email"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              required
+            />
+          </div>
+          <div>
+            <Label
+              htmlFor="password"
+              className="mb-1 block text-sm font-medium"
+            >
+              Password
+            </Label>
+            <TextInput
+              id="password"
+              type="password"
+              placeholder="Enter your password"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              required
+            />
+          </div>
+          <Button type="submit" className="w-full">
+            Sign In
+          </Button>
+        </form>
+        <p className="mt-4 text-center text-sm text-gray-600 dark:text-gray-300">
+          Don't have an account?{" "}
+          <a href="/auth/signup" className="text-blue-500 hover:underline">
+            Sign up
+          </a>
+        </p>
+      </Card>
+    </div>
+  );
+};
