+++ nextjs-task-system/app/auth/signin/page.tsx
@@ -1,90 +1,27 @@
-"use client";
+import { Alert } from "flowbite-react";
+import { SignIn } from "./SignIn";
+import { getUsers } from "@/app/actions/getUsers";
 
-import { signIn } from "next-auth/react";
-import { useState } from "react";
-import { Button, TextInput, Alert, Card, Label } from "flowbite-react";
-import { useRouter } from "next/navigation";
+const ENABLE_DEMO = true;
 
-const SignInPage = () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState("");
-  const router = useRouter();
-
-  const handleSignIn = async (e: React.FormEvent) => {
-    try {
-      e.preventDefault();
-      setError("");
-      const result = await signIn("credentials", {
-        email,
-        password,
-        role: "Admin",
-        redirect: false,
-      });
-
-      if (result?.error) {
-        setError("Invalid credentials");
-      } else {
-        router.push("/");
-      }
-    } catch (error) {
-      setError("unexpected error");
-    }
-  };
-
+const SignInPage = async () => {
+  const users = ENABLE_DEMO ? await getUsers() : [];
   return (
-    <div className="flex flex-1 items-center justify-center">
-      <Card className="w-full max-w-md rounded-lg p-6 shadow-md">
-        <h1 className="mb-6 text-center text-2xl font-bold dark:text-gray-200">
-          Sign In
-        </h1>
-        {error && (
-          <Alert color="failure" className="mb-4">
-            {error}
-          </Alert>
-        )}
-        <form onSubmit={handleSignIn} className="space-y-4">
-          <div>
-            <Label htmlFor="email" className="mb-1 block text-sm font-medium">
-              Email
-            </Label>
-            <TextInput
-              id="email"
-              type="email"
-              placeholder="Enter your email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              required
-            />
-          </div>
-          <div>
-            <Label
-              htmlFor="password"
-              className="mb-1 block text-sm font-medium"
-            >
-              Password
-            </Label>
-            <TextInput
-              id="password"
-              type="password"
-              placeholder="Enter your password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              required
-            />
-          </div>
-          <Button type="submit" className="w-full">
-            Sign In
-          </Button>
-        </form>
-        <p className="mt-4 text-center text-sm text-gray-600 dark:text-gray-300">
-          Don't have an account?{" "}
-          <a href="/auth/signup" className="text-blue-500 hover:underline">
-            Sign up
-          </a>
-        </p>
-      </Card>
-    </div>
+    <SignIn>
+      {ENABLE_DEMO && (
+        <Alert color="success">
+          <span className="font-medium">Users demo</span>
+          {users.map((user) => (
+            <p>
+              <span className="mr-5">{user.email}</span>
+              <span className="ml-auto">
+                password: {user.role === "Admin" ? "admin" : "user"}
+              </span>
+            </p>
+          ))}
+        </Alert>
+      )}
+    </SignIn>
   );
 };
